Update rendering flow
